Feat: Update, Delete 비밀번호 확인 구현

+++ src/main/java/com/board/board/Service/PostService.java
@@ -1,5 +1,6 @@
 package com.board.board.Service;
 
+import com.board.board.Dto.PasswordDto;
 import com.board.board.Dto.PostRequestDto;
 import com.board.board.Entity.Post;
 import com.board.board.Repository.PostRepository;
@@ -19,7 +20,8 @@ public class PostService {
 
     public PostResponseDto createPost(PostRequestDto postRequestDto) {
         Post post = new Post(postRequestDto);
-        Post savePost = postRepository.save(post);
+
+        postRepository.save(post);
 
         PostResponseDto postResponseDto = new PostResponseDto(post);
         return postResponseDto;
@@ -32,23 +34,37 @@ public class PostService {
 
     @Transactional
     public Long putPost(Long id, PostRequestDto requestDto) {
-        //해당 메모가 DB에 존재하는지 확인
-        Post post = findPost(id);
+//          해당 메모가 DB에 존재하는지 확인
+            Post post = findPost(id);
 
-        post.update(requestDto);
+            // 비밀번호 확인
+            if(post.getPW().equals(requestDto.getPassword())) {
+                post.update(requestDto);
+            }
+            else{
+                throw new IllegalArgumentException("비밀번호가 틀립니다");
+            }
+            return id;
+       }
 
-        return id;
-    }
+    public Long deletePost(Long id, PasswordDto passwordDto) {
+        // 해당 메모가 DB에 존재하는지 확인
+            Post post = findPost(id);
 
-    public Long deletePost(Long id) {
-        Post post = findPost(id);
+        // 비밀번호 확인
+            if(post.getPW().equals(passwordDto.getPassword())) {
+                postRepository.delete(post);
+            }
+            else{
+                throw new IllegalArgumentException("비밀번호가 올바르지 않습니다.");
+            }
+            return id;
+        }
 
-        postRepository.delete(post);
-        return id;
-    }
+
+
     private Post findPost(Long id){
         return postRepository.findById(id).orElseThrow(() ->
                 new RuntimeException("선택한 게시글이 존재하지 않습니다."));
     }
-
 }
